Restructed tests and moved config file to a separate folder to make it available for import
