+++ backend-docker/server/apiserver_test.go
@@ -3,16 +3,10 @@ package server
 import (
 	"testing"
 	"github.com/EUDAT-GEF/GEF/backend-docker/dckr"
-
-
 	"net/http/httptest"
-
 	"fmt"
 	"net/http"
-
 )
-
-
 
 func TestServer(t *testing.T) {
 	c := newClient(t)
@@ -32,7 +26,7 @@ func TestServer(t *testing.T) {
 	if res.StatusCode != 200 {
 		t.Errorf("Success expected: %d", res.StatusCode)
 	}
-	
+
 }
 
 func createServer(client dckr.Client, t *testing.T) *Server {
